修正新增、刪除、修改 錯誤處理

+++ routes/posts.js
@@ -14,10 +14,18 @@ router.get('/', async (req, res, next) =>{
 // 新增一筆貼文
 router.post('/', async(req, res, next) =>{
     try {
-        const { name, content, image, like, tags } = req.body;
+        const { name, content, image, likes, tags } = req.body;
         if (content && name && tags.length) {
-            const newPost = await Post.create(req.body)
-            handleSuccess(res, newPost, '新增成功')
+            const newPost = await Post.create(
+                { 
+                    name: typeof name === 'string' ? name.trim() : "",  
+                    content: typeof content === 'string' ? content.trim() : "", 
+                    image: typeof image === 'string' ? image.trim() : "",
+                    likes,
+                    tags
+                }
+            )
+            handleSuccess(res, newPost)
         } else {
             handleError(res);
         }
@@ -28,26 +36,56 @@ router.post('/', async(req, res, next) =>{
 
 // 刪除所有貼文
 router.delete('/', async(req, res, next) =>{
-    const data = await Post.deleteMany({});
-    handleSuccess(res, data, '修改成功')
+    if(req.originalUrl === '/posts') {
+        const data = await Post.deleteMany({});
+        handleSuccess(res, data, '修改成功')
+    } else {
+        handleError(res);
+    }
 });
 
 // 刪除一筆貼文
 router.delete('/:id', async(req, res, next) =>{
-    const id = req.params.id
-    await Post.findByIdAndDelete(id);
-    handleSuccess(res, null, '刪除成功')
+    try {
+        const id = req.params.id
+        const findId = await Post.findById(id)
+        if(findId.id) {
+            const deletePost = await Post.findByIdAndDelete(id, {},{
+                new: true,
+            })
+            handleSuccess(res, deletePost, '刪除成功');
+        } else {
+            handleError(res);
+        }
+    } catch (error) {
+        handleError(res, error);
+    }
 });
 
 // 更新一筆貼文
 router.patch('/:id', async(req, res, next) =>{
-    const id = req.params.id
-    const { name, content, tags, likes, image } = req.body
-    if (id && (name || content || tags.length)) {
-        await Post.findByIdAndUpdate(id, { name, content, tags, likes, image  });
-        handleSuccess(res, null, '更新成功')
-    } else {
-        handleError(res);
+    try {
+        const id = req.params.id
+        const findId = await Post.findById(id)
+        const { name, content, tags, likes, image } = req.body
+        if (findId.id && name && content && tags.length) {
+            const updatePost = await Post.findByIdAndUpdate(id, {
+                name: typeof name === 'string' ? name.trim() : "",  
+                content: typeof content === 'string' ? content.trim() : "", 
+                image: typeof image === 'string' ? image.trim() : "",
+                likes, 
+                tags
+              },
+              {
+                new: true,
+                runValidators: true
+              });
+            handleSuccess(res, updatePost, '更新成功')
+        } else {
+            handleError(res);
+        }
+    } catch (error) {
+        handleError(res, error);
     }
 });
